+++ core/src/main/java/org/infinispan/marshall/core/impl/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * @private
+ */
+package org.infinispan.marshall.core.impl;
